use qint64 for the document size
svn path=/trunk/KDE/kdegraphics/okular/; revision=719107

+++ core/document.cpp
@@ -1045,7 +1045,7 @@ bool Document::openDocument( const QString & docFile, const KUrl& url, const KMi
 {
     KMimeType::Ptr mime = _mime;
     QByteArray filedata;
-    int document_size = -1;
+    qint64 document_size = -1;
     bool isstdin = url.fileName( KUrl::ObeyTrailingSlash ) == QLatin1String( "-" );
     if ( !isstdin )
     {
